Support passing raw kernel arguments
User-defined types (e.g., `struct` or `class`) are currently not
supported as kernel arguments by DPCTL. As the C++ type information for
the arguments is not available in DPCTL, this implementation relies on
the `raw_kernel_arg` experimental extension to allow users to pass
arbitrary arguments.

It's the user's responsibility to pack and layout the argument as
expected by kernel ABI and obtain a suitable pointer to construct the
`raw_kernel_arg`.

`raw_kernel_arg` is non-owning, so it's the user's responsibility to
keep the underlying data alive.

Signed-off-by: Lukas Sommer <lukas.sommer@codeplay.com>